classes moved to classes folder and json methods applied to person and rentals
