+++ classes/capitalizedecorator.rb
