+++ classes/decorator.rb
