+++ classes/person.rb
@@ -1,0 +1,49 @@
+require_relative './nameable'
+require_relative './rental'
+require 'json'
+
+class Person < Nameable
+  attr_accessor :name, :age, :parent_permission
+  attr_reader :id, :rentals
+
+  def initialize(age, name = 'Unknown', parent_permission: true, id = Random.rand(1..10_000))
+    super()
+    @id = id
+    @name = name
+    @age = age
+    @parent_permission = parent_permission
+    @rentals = []
+  end
+
+  def can_use_services?
+    of_age? || @parent_permission
+  end
+
+  def add_rental(book, date)
+    Rental.new(date, book, self)
+  end
+
+  def correct_name
+    @name
+  end
+
+  private
+
+  def of_age?
+    @age >= 18
+  end
+
+  def to_json(*_args)
+    {
+      JSON.create_id => self.class.name,
+      'id' => @id,
+      'name' => @name,
+      'age' => @age,
+      'parent_permission' => @parent_permission,
+    }.to_json(*_args)
+  end
+
+  def self.json_create(object)
+    new(object['age'], object['name'], parent_permission: object['parent_permission'])
+  end
+end
