+++ classes/trimmerdecorator.rb
