--- person.rb
@@ -1,34 +1,0 @@
-require_relative './nameable'
-require_relative './rental'
-
-class Person < Nameable
-  attr_accessor :name, :age
-  attr_reader :id, :rentals
-
-  def initialize(age, name = 'Unknown', parent_permission: true)
-    super()
-    @id = Random.rand(1..10_000)
-    @name = name
-    @age = age
-    @parent_permission = parent_permission
-    @rentals = []
-  end
-
-  def can_use_services?
-    of_age? || @parent_permission
-  end
-
-  def add_rental(book, date)
-    Rental.new(date, book, self)
-  end
-
-  def correct_name
-    @name
-  end
-
-  private
-
-  def of_age?
-    @age >= 18
-  end
-end
